Update so the id for agenda and agenda items is set using UUID during creation in the app.
This was done to make sure we don't run into any equality issues with objects before they are saved into the db. Since not a ton should be saved and only at the pace of a single user, creating a sequence in the db seemed overkill. If had high throughput, would use a sequence in the db to allow horizontal scaling of applications.